+++ src/components/lobby/LogoutIcon.js
@@ -1,21 +1,49 @@
 import React from 'react';
 import Tooltip from '@material-ui/core/Tooltip';
 import ExitToAppIcon from "@material-ui/icons/ExitToApp";
+import { withRouter } from "react-router-dom";
+import Button from "@material-ui/core/Button";
+//Redux
+import { connect } from "react-redux";
+import { logoutUser } from "../../redux/actions/userActions";
 
-export default function LogoutIcon() {
-    const [open, setOpen] = React.useState(false);
+class LogoutIcon extends React.Component {
 
-    const handleTooltipClose = () => {
-        setOpen(false);
+    constructor() {
+        super();
+        this.state = {
+            open: false,
+        };
+    }
+
+    async logout() {
+        await this.props.logoutUser();
+        //await this.props.resetState();
+        this.props.history.push("/login");
+    }
+
+    handleTooltipClose = () => {
+        this.setState({open: false});
     };
 
-    const handleTooltipOpen = () => {
-        setOpen(true);
+    handleTooltipOpen = () => {
+        this.setState({open: true});
     };
 
-    return (
-            <Tooltip disableFocusListener disableTouchListener title="Logout">
-                <ExitToAppIcon style = {{ display: "inline", fontSize: 30, position: "absolute", right: 20, top: 20}}/>
-            </Tooltip>
-    );
+    render() {
+        return (
+            <Button
+                style={{size: "fit-content", display: "inline", fontSize: 30, position: "absolute", right: 20, top: 20}}
+                onClick={() => {
+                    this.logout();
+                }}>
+                <Tooltip disableFocusListener disableTouchListener title="Logout">
+                    <ExitToAppIcon style={{fontSize: 30}}/>
+                </Tooltip>
+            </Button>
+
+        );
+    }
 }
+
+export default withRouter(connect(null, { logoutUser })(LogoutIcon));
